# format: Reduce allocations made during formatting