--- a/format_test.go
+++ b/format_test.go
@@ -191,3 +191,71 @@ func TestDecimalString(t *testing.T) {
 		}
 	}
 }
+
+type devNull struct{}
+
+func (d devNull) Write(b []byte) (int, error) {
+	return len(b), nil
+}
+
+// BenchmarkFormat benchmarks formatting using the various format specifiers.
+func BenchmarkFormat(b *testing.B) {
+	tests := []struct {
+		name string
+		txt  string
+		fmt  string
+		want string
+	}{{
+		name: "large number f format",
+		txt:  "12345678901234.67890123456789012345",
+		fmt:  "%f",
+		want: "12345678901234.678901",
+	}, {
+		name: "large number e format",
+		txt:  "12345678901234.67890123456789012345",
+		fmt:  "%e",
+		want: "1.234568e+13",
+	}, {
+		name: "large number padding f format",
+		txt:  "12345678901234.67890123456789012345",
+		fmt:  "%80.40f",
+		want: "                         12345678901234.6789012345678901234500000000000000000000",
+	}, {
+		name: "large number padding e format",
+		txt:  "12345678901234.67890123456789012345",
+		fmt:  "%80.40e",
+		want: "                                  1.2345678901234678901234567890123450000000e+13",
+	}, {
+		name: "special",
+		txt:  "nan",
+		fmt:  "%f",
+		want: "NaN",
+	}, {
+		name: "special with padding",
+		txt:  "nan",
+		fmt:  "%80.40f",
+		want: "                                                                             NaN",
+	}}
+
+	for _, tc := range tests {
+		tc := tc
+		b.Run(tc.name, func(b *testing.B) {
+			// Ensure correctness of test case before benchmarking.
+			v := MustParse(tc.txt)
+			vptr := &v // Avoid allocating during Fprintf() call
+			got := fmt.Sprintf(tc.fmt, vptr)
+			if got != tc.want {
+				b.Fatalf("Unexpected formatted value. got %s, "+
+					"want %s", got, tc.want)
+			}
+			w := devNull{}
+
+			// Benchmark.
+			b.ResetTimer()
+			b.ReportAllocs()
+			for i := 0; i < b.N; i++ {
+				fmt.Fprintf(w, tc.fmt, vptr)
+			}
+		})
+	}
+}