Reducing the number of allocations is an important performance improvement, as it reduces the pressure for GC events. This PR is split into various commits to ease reviewing. The first one adds a benchmark against which to compare the improvements, and then each following PR makes incremental changes, in order to make it clear they don't change the semantics for formatting.

Comparison of the benchmarks before/after this set of commits:

```
name                                    old time/op    new time/op    delta
Format/large_number_f_format-4             688ns ± 1%     609ns ± 1%  -11.54%  (p=0.000 n=10+10)
Format/large_number_e_format-4             686ns ± 1%     608ns ± 1%  -11.49%  (p=0.000 n=10+10)
Format/large_number_padding_f_format-4     882ns ± 1%     732ns ± 1%  -17.06%  (p=0.000 n=10+10)
Format/large_number_padding_e_format-4     840ns ± 1%     723ns ± 2%  -13.99%  (p=0.000 n=10+8)
Format/special-4                           130ns ± 2%     118ns ± 3%   -8.85%  (p=0.000 n=10+10)
Format/special_with_padding-4              261ns ± 1%     273ns ±14%     ~     (p=0.079 n=9+10)

name                                    old alloc/op   new alloc/op   delta
Format/large_number_f_format-4              112B ± 0%       24B ± 0%  -78.57%  (p=0.000 n=10+10)
Format/large_number_e_format-4             88.0B ± 0%     16.0B ± 0%  -81.82%  (p=0.000 n=10+10)
Format/large_number_padding_f_format-4      304B ± 0%       80B ± 0%  -73.68%  (p=0.000 n=10+10)
Format/large_number_padding_e_format-4      200B ± 0%       80B ± 0%  -60.00%  (p=0.000 n=10+10)
Format/special-4                           0.00B          0.00B          ~     (all equal)
Format/special_with_padding-4              80.0B ± 0%     80.0B ± 0%     ~     (all equal)

name                                    old allocs/op  new allocs/op  delta
Format/large_number_f_format-4              3.00 ± 0%      1.00 ± 0%  -66.67%  (p=0.000 n=10+10)
Format/large_number_e_format-4              3.00 ± 0%      1.00 ± 0%  -66.67%  (p=0.000 n=10+10)
Format/large_number_padding_f_format-4      5.00 ± 0%      1.00 ± 0%  -80.00%  (p=0.000 n=10+10)
Format/large_number_padding_e_format-4      4.00 ± 0%      1.00 ± 0%  -75.00%  (p=0.000 n=10+10)
Format/special-4                            0.00           0.00          ~     (all equal)
Format/special_with_padding-4               1.00 ± 0%      1.00 ± 0%     ~     (all equal)
```